encode value types (again) (#29)
* remove GREMLIN_TINKERGRAPH_ONDISK_ROOT_DIR config option

1) it's overlapping in meaning with GREMLIN_TINKERGRAPH_GRAPH_LOCATION
2) it can be controlled on startup from env/java vars anyway

* bring value type encoding back, this time a little cleaner

when serializing properties we need to encode the id type in a
separate entry, to ensure we can deserialize it back to the very same
type. I would have hoped that MsgPack does that for us, but that's
only partly the case.

E.g. the different integer types cannot be distinguished other than by their value. When we deserialize `42`, we have no idea whether it should be deserialized as a byte, short, integer or double.

* fix list property serialization

* logging messages: starting with existing storage; completed in XYZms

silly useless java lambdas...