Edge properties are also stored for direction IN refs.
We also handle the case of multiple edges with the same label between
the same node pair.

OverflowDbEdges are invalid after another edge was added to/removed from an
adjacent vertex because the stored block offsets can change in this
case. We need to implement a mechanism to at least detect
this.